rename const memo to transfer_memo
