+++ programs/whirlpool/src/constants/mod.rs
@@ -1,5 +1,5 @@
 pub mod nft;
-pub mod memo;
+pub mod transfer_memo;
 pub mod test_constants;
 
 pub use nft::*;
