+++ programs/whirlpool/src/instructions/v2/collect_protocol_fees.rs
@@ -1,4 +1,4 @@
-use crate::{constants::memo, state::*, util::v2::transfer_from_vault_to_owner_v2};
+use crate::{constants::transfer_memo, state::*, util::v2::transfer_from_vault_to_owner_v2};
 use anchor_lang::prelude::*;
 use anchor_spl::token_interface::{Mint, TokenAccount, TokenInterface};
 use anchor_spl::memo::Memo;
@@ -46,7 +46,7 @@ pub fn handler(ctx: Context<CollectProtocolFeesV2>) -> Result<()> {
         &ctx.accounts.token_program_a,
         &ctx.accounts.memo_program,
         whirlpool.protocol_fee_owed_a,
-        memo::TRANSFER_MEMO_COLLECT_PROTOCOL_FEES.as_bytes(),
+        transfer_memo::TRANSFER_MEMO_COLLECT_PROTOCOL_FEES.as_bytes(),
     )?;
 
     transfer_from_vault_to_owner_v2(
@@ -57,7 +57,7 @@ pub fn handler(ctx: Context<CollectProtocolFeesV2>) -> Result<()> {
         &ctx.accounts.token_program_b,
         &ctx.accounts.memo_program,
         whirlpool.protocol_fee_owed_b,
-        memo::TRANSFER_MEMO_COLLECT_PROTOCOL_FEES.as_bytes(),
+        transfer_memo::TRANSFER_MEMO_COLLECT_PROTOCOL_FEES.as_bytes(),
     )?;
 
     Ok(ctx.accounts.whirlpool.reset_protocol_fees_owed())
